更新百度去广告模块

+++ app/src/main/java/com/beforecar/ad/utils/FileUtils.java
@@ -18,6 +18,7 @@ public class FileUtils {
      */
     public static boolean delete(final File file) {
         if (file == null) return false;
+        if (!file.exists()) return true;
         if (file.isDirectory()) {
             return deleteDir(file);
         }
